[DNM] manual coala error fixes
Do not merge.

These changes are to fix coala errors.

The Markdown files are reformatted due to consistency problem, because
of that the options that was defined can be removed.

SpaceConsistencyBear is added to Markdown to prevent trailing
whitespaces.

The XML files are reformatted with XMLBear and since XMLBear can't be
configured that much (especially for things like word wrapping and
breaking down XML syntaxes to multiple lines)